Move BackendHttpClient to the test project
